requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
djangorestframework-simplejwt==5.3.0 
psycopg2-binary==2.9.9 
redis==5.0.1 
channels==4.0.0 
channels-redis==4.1.0 
celery==5.3.4 
Pillow==10.1.0 
drf-yasg==1.21.7 
python-dotenv==1.0.0 
pandas==2.0.3 
numpy==1.24.3 
scikit-learn==1.3.0 
joblib==1.3.2 
gunicorn==21.2.0 
django-filter==23.5 
django-allauth==0.61.1 
django-import-export==3.3.5 
django-redis==5.4.0 
websockets==12.0 
django-model-utils==4.3.1 
django-storages==1.14.2 
boto3==1.34.14 
django-anymail==10.2 
django-ckeditor==6.7.0 
django-mptt==0.14.0 
django-taggit==4.0.0 
django-ratelimit==4.1.0 
django-axes==6.0.1 
pyotp==2.9.0 
qrcode==7.4.2 
